requirements.txt
# Core
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
gunicorn>=21.0

# ML — Traditional
scikit-learn>=1.3
joblib>=1.3
numpy>=1.24

# ML — Transformer (DistilBERT)
transformers>=4.35
torch>=2.0

# AI Generation
google-generativeai>=0.8

# Document Formatting
python-docx>=1.1
reportlab>=4.1
markdown>=3.5

# Research / Evaluation
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12

# Authentication
flask-bcrypt>=1.0.1
PyJWT>=2.8.0
google-auth>=2.27.0
requests>=2.31.0
